Tweak CPU requests to match reality
We can help shrink the number of Nodes needed in the Kubernetes cluster
by reducing the CPU requests to match reality. This service uses almost
no CPU regularly, and so the requests (guaranteed CPU allocation) can be
reduced. The limits remain the same, so that the service can burst to
high CPU usage if needed.

This will help reduce costs in the Kubernetes cluster by reducing the
number of Nodes (EC2 VMs) needed to run the workload.